+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 - The keywords checklist now knows what spaces are and count words accordingly
 - Fixed keywords checklist not working after live preview is opened
 - Fixed bug where fallback social images would cause SEO field to error
+- Keyword density now works correctly for keywords containing more than one word
+- Fixed a bug where deleting a keyword would occasionally cause a JS error 
 
 ## 3.1.0 - 2018-03-02
 ### Added